backend/auth_service/requirements.txt
fastapi==0.100.0
pydantic==2.1.1
uvicorn==0.22.0
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
asyncpg==0.28.0
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.2
pydantic-settings==2.0.2
httpx==0.28.1
PyJWT==2.3.0
fastapi-mail 